add search filtering to the table
using alpine js. I'm not super happy with this solution because
the logic is separated into 3 different template files. If search
field will be part of the nav header how should it integrate?

Maybe search only makes sense on the table and I should move all this
filtering logic into the table's template.